modularise two database queries, relates #146
